feat(main): initialize image_resizer
